ZipCode/requirements.txt
astroid==1.6.5
Django==2.0.7
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
pylint==1.9.2
pytz==2018.5
six==1.11.0
wrapt==1.10.11
